Restore the swept-face example, make an independent cable bracket example
